+++ src/main/scala/api/payload/request/PrivacyRequestPayload.scala
@@ -1,5 +1,5 @@
 package io.blindnet.privacy
-package api.endpoints.payload
+package api.endpoints.payload.request
 
 import cats.effect.*
 import cats.implicits.*
